Crate mpid_messaging

MPID Messaging

Types used by SAFE Vaults and SAFE Clients to send and receive messages.

Structs

Header
A simple struct holding basic info of the message

MpidHeader
To be used as a notification to the receiver

MpidMessage
Full message sending/fetching to/from the Network

Enums

MpidMessageWrapper
Wrapper of a notification or a full message