Enum mpid_messaging::MpidMessageWrapper

pub enum MpidMessageWrapper {
    Online,
    PutMessage(MpidMessage),
    PutHeader(MpidHeader),
    GetMessage(MpidHeader),
    OutboxHas(Vec<XorName>),
    OutboxHasResponse(Vec<MpidHeader>),
    GetOutboxHeaders,
    GetOutboxHeadersResponse(Vec<MpidHeader>),
    DeleteMessage(XorName),
    DeleteHeader(XorName),
}

A serialisable wrapper to allow multiplexing all MPID message types and actions via a single type.

Variants

Online

Sent by a Client to its MpidManagers to notify them that it has just connected to the network.

PutMessage

Sent by a Client to its MpidManagers when storing a new MpidMessage.

PutHeader

Sent by the sender's MpidManagers to the receiver's MpidManagers to alert them of a new message.

GetMessage

Sent by the receiver to its MpidManagers to try to retrieve the message corresponding to the header.

OutboxHas

Sent by a Client to its MpidManagers to query whether the provided vector of message names continue to exist as messages in its outbox.

OutboxHasResponse

Sent by MpidManagers to the Client as a response to an OutboxHas. The contents is a subset of the list provided in the corresponding OutboxHas.

GetOutboxHeaders

Sent by a Client to its MpidManagers to retrieve the list of headers of all messages in its outbox.

GetOutboxHeadersResponse

Sent by MpidManagers to the Client as a response to a GetOutboxHeaders. The contents is the list of headers of all messages in the outbox.

DeleteMessage

Sent by a Client to its MpidManagers to delete the named message from its inbox or outbox.

DeleteHeader

Sent by a receiving Client to the sender's MpidManagers to delete the named message's header from the sender's outbox.