Struct mpid_messaging::MpidMessage

pub struct MpidMessage {
    // some fields omitted
}

A full message including header and body which can be sent to or retrieved from the network.

Methods

impl MpidMessage

fn new(sender: XorName, metadata: Vec<u8>, recipient: XorName, body: Vec<u8>, secret_key: &SecretKey) -> Result<MpidMessage, Error>

Constructor.

sender and metadata are used to construct an MpidHeader member, accessed via the header() getter. For details on these arguments, see MpidHeader::new().

recipient represents the name of the intended receiver of the message.

body is arbitrary, user-supplied data representing the main portion of the message. It must not exceed MAX_BODY_SIZE. It can be empty if desired.

An error will be returned if body exceeds MAX_BODY_SIZE, if MpidHeader::new() fails or if serialisation during the signing process fails.

fn header(&self) -> &MpidHeader

Getter for MpidHeader member, created when calling new().

fn recipient(&self) -> &XorName

The name of the intended receiver of the message.

fn body(&self) -> &Vec<u8>

Arbitrary, user-supplied data representing the main portion of the message.

fn name(&self) -> Result<XorName, Error>

The name of the message, equivalent to the MpidHeader::name(). As per that getter, this is relatively expensive, so its use should be minimised.

fn verify(&self, public_key: &PublicKey) -> bool

Validates the message and header signatures against the provided PublicKey.