Crate mpid_messaging

MPID Messaging

Types used by SAFE Vaults and SAFE Clients to send and receive messages.

Structs

MpidHeader
Minimal information about a given message which can be used as a notification to the receiver.

MpidMessage
A full message including header and body which can be sent to or retrieved from the network.

Enums

Error
Error types relating to MPID messaging.

MpidMessageWrapper
A serialisable wrapper to allow multiplexing all MPID message types and actions via a single type.

Constants

GUID_SIZE
Length of the GUID (unique ID) of a message (16 bytes).

MAX_BODY_SIZE
Maximum allowed length for a message's body (101,760 bytes).

MAX_HEADER_METADATA_SIZE
Maximum allowed length for a header's metadata (128 bytes).

MAX_INBOX_SIZE
Maximum allowed inbox size for an account (128 MiB).

MAX_OUTBOX_SIZE
Maximum allowed outbox size for an account (128 MiB).